Some refactoring and struct unification around the upload queue
